Reorganize tach code and hide tach's default result tables
